use logical instead of physical location (#5)

+++ crates/sonate/src/lib.rs
@@ -107,7 +107,7 @@ impl Engine {
                 }
             }),
             on_resize: Box::new(move |width, height| {
-                let _ = resize_sender.send(Command::SetViewportSize(width as f64, height as f64));
+                let _ = resize_sender.send(Command::SetViewportSize(width, height));
             }),
         };
